Some major reworking of the Gruntfile
Still more work to be done, the final build process doesn't quite work
yet.  The dev part of it works better now (wrapped the less task with
newer and cleaned old tasks up).  Will need to make some other tasks as
well, something to seed the database with data for testing, and
something to deploy the app.

Can work on converting the client side JS to Coffee now too.